fix: 무한 렌더되는 이슈 해결
